Downgraded Stackage snapshot to lts-21.25
This was required to make this project build on older Ubuntu versions
for preparing the AppImage distribution.

A consequence of this is we also had to downgrade from the modern
prettyprinter library to the older ansi-wl-pprint library which is
deprecated at a later time. These libraries are used by the
MetarDisplay.Opts module for usage text formatting.